SCRUM-63 refactor code - move CarListResponse.kt to ws module - delete unneeded interfaces
