+++ app/src/main/java/com/air_wheelly/wheelly/domain/car_list/CarListModel.kt
@@ -1,7 +1,7 @@
 package com.air_wheelly.wheelly.domain.car_list
 
-import hr.air_wheelly.core.network.CarListResponse
 import hr.air_wheelly.core.util.EnumFuelType
+import hr.air_wheelly.ws.models.responses.CarListResponse
 
 class CarListModel {
     fun filterCars(
@@ -12,7 +12,7 @@ class CarListModel {
     ): List<CarListResponse> {
         return carList.filter { car ->
             (fuelTypes.isEmpty() || fuelTypes.contains(EnumFuelType.valueOf(car.fuelType!!.uppercase()))) &&
-                    (manufacturer.isEmpty() || car.model?.manafacturerId?.contains(manufacturer, ignoreCase = true) == true) &&
+                    (manufacturer.isEmpty() || car.model?.manufacturerId?.contains(manufacturer, ignoreCase = true) == true) &&
                     (year == null || car.yearOfProduction == year)
         }
     }
