+++ app/src/main/java/com/air_wheelly/wheelly/domain/payment/PaymentViewModel.kt
@@ -6,10 +6,10 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.braintreepayments.api.DropInClient
 import com.braintreepayments.api.DropInRequest
-import hr.air_wheelly.core.network.CarListResponse
 import hr.air_wheelly.core.network.ResponseListener
 import hr.air_wheelly.core.network.models.ErrorResponseBody
 import hr.air_wheelly.core.network.models.SuccessfulResponseBody
+import hr.air_wheelly.ws.models.responses.CarListResponse
 import hr.air_wheelly.ws.request_handlers.CarByIdRequestHandler
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
