--- core/src/main/java/hr/air_wheelly/core/network/CarHandler.kt
@@ -1,5 +1,0 @@
-package hr.air_wheelly.core.network
-
-interface CarHandler {
-    suspend fun getCarListings(carListOutcomeListener: CarListOutcomeListener)
-}
