--- core/src/main/java/hr/air_wheelly/core/network/CarListOutcomeListener.kt
@@ -1,6 +1,0 @@
-package hr.air_wheelly.core.network
-
-interface CarListOutcomeListener {
-    fun onSuccessfulCarListFetch(response: List<CarListResponse>)
-    fun onFailedCarListFetch(reason: String)
-}
