+++ ws/src/main/java/hr/air_wheelly/ws/request_handlers/CarListRequestHandler.kt
@@ -1,10 +1,8 @@
 package hr.air_wheelly.ws.request_handlers
 
 import android.content.Context
-import hr.air_wheelly.core.network.CarListResponse
 import hr.air_wheelly.core.network.models.SuccessfulResponseBody
-import hr.air_wheelly.ws.models.responses.CarLocationResponse
-import hr.air_wheelly.ws.models.responses.car.CarLocationBody
+import hr.air_wheelly.ws.models.responses.CarListResponse
 import hr.air_wheelly.ws.network.NetworkService
 import retrofit2.Call
